app/requirements.txt
influxdb==5.3.1
python-dotenv==1.0.0
requests==2.31.0
